requirements.txt
streamlit==1.33.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-timeout
pytest-cov
pylint
